% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_type.R
\name{get_data_type}
\alias{get_data_type}
\title{Determine Input Data Type for Severity Pipelinele}
\usage{
get_data_type(data = NULL, total_count = NULL, total_death = NULL)
}
\arguments{
\item{data}{A dataframe, incidence object, or NULL.}

\item{total_count}{Numeric total cases (optional).}

\item{total_death}{Numeric total deaths (optional).}
}
\value{
One of "count_data", "linelist", "incidence".
}
\description{
Determine Input Data Type for Severity Pipelinele
}
